Display window that lets user to set daemon binary path if syncthing binary is not found, but syncthing configuration exists

+++ syncthing_gtk/daemonprocess.py
@@ -22,6 +22,8 @@ class DaemonProcess(GObject.GObject):
 		b"line"			: (GObject.SIGNAL_RUN_FIRST, None, (object,)),
 		# exit(code)	- emited when process exits
 		b"exit"			: (GObject.SIGNAL_RUN_FIRST, None, (int,)),
+		# failed(exception) - emited if process fails to start
+		b"failed"		: (GObject.SIGNAL_RUN_FIRST, None, (object,)),
 	}
 	SCROLLBACK_SIZE = 500	# Maximum number of output lines stored in memory
 	
@@ -29,21 +31,27 @@ class DaemonProcess(GObject.GObject):
 		""" commandline should be list of arguments """
 		GObject.GObject.__init__(self)
 		self.commandline = commandline
+		self._proc = None
 	
 	def start(self):
 		os.environ["STNORESTART"] = "1"	# see syncthing --help
-		self._cancel = Gio.Cancellable()
-		if HAS_SUBPROCESS:
-			# Unix
-			flags = Gio.SubprocessFlags.STDOUT_PIPE | Gio.SubprocessFlags.STDERR_MERGE
-			self._proc = Gio.Subprocess.new(commandline, flags)
-			self._proc.wait_check_async(None, self._cb_finished)
-			self._stdout = self._proc.get_stdout_pipe()
-		else:
-			# Gio < 3.12 - Gio.Subprocess is missing :(
-			self._proc = Popen(commandline, stdout=PIPE)
-			self._stdout = Gio.UnixInputStream.new(self._proc.stdout.fileno(), False)
-			self._check = GLib.timeout_add_seconds(1, self._cb_check_alive)
+		try:
+			self._cancel = Gio.Cancellable()
+			if HAS_SUBPROCESS:
+				# Unix
+				flags = Gio.SubprocessFlags.STDOUT_PIPE | Gio.SubprocessFlags.STDERR_MERGE
+				self._proc = Gio.Subprocess.new(self.commandline, flags)
+				self._proc.wait_check_async(None, self._cb_finished)
+				self._stdout = self._proc.get_stdout_pipe()
+			else:
+				# Gio < 3.12 - Gio.Subprocess is missing :(
+				self._proc = Popen(self.commandline, stdout=PIPE)
+				self._stdout = Gio.UnixInputStream.new(self._proc.stdout.fileno(), False)
+				self._check = GLib.timeout_add_seconds(1, self._cb_check_alive)
+		except Exception, e:
+			# Startup failed
+			self.emit("failed", e)
+			return
 		self._lines = deque([], DaemonProcess.SCROLLBACK_SIZE)
 		self._buffer = ""
 		self._stdout.read_bytes_async(256, 0, self._cancel, self._cb_read, ())
@@ -110,21 +118,23 @@ class DaemonProcess(GObject.GObject):
 	
 	def terminate(self):
 		""" Terminates process (sends SIGTERM) """
-		if HAS_SUBPROCESS:
-			# Gio.Subprocess
-			self._proc.send_signal(15)
-		else:
-			# subprocess.Popen
-			self._proc.terminate()
+		if not self._proc is None:
+			if HAS_SUBPROCESS:
+				# Gio.Subprocess
+				self._proc.send_signal(15)
+			else:
+				# subprocess.Popen
+				self._proc.terminate()
 	
 	def kill(self):
 		""" Kills process (sends SIGTERM) """
-		if HAS_SUBPROCESS:
-			# Gio.Subprocess
-			self._proc.force_exit()
-		else:
-			# subprocess.Popen
-			self._proc.kill()
+		if not self._proc is None:
+			if HAS_SUBPROCESS:
+				# Gio.Subprocess
+				self._proc.force_exit()
+			else:
+				# subprocess.Popen
+				self._proc.kill()
 	
 	def get_output(self):
 		""" Returns process output as iterable list of lines """
